Script that adds dynamicbone to array of objects.
This script adds the dynamic-bone script to an array of objects. Contains all the parameters that the dynamic bone script includes.

How to use: 
Populate the Array at the top of the script with the gameobjects you want to have the dynamic bone script.
Double-check all your settings and check the box at the bottom that says "Check When Ready".
Right click the script-name in your inspector-panel and select "Click here to add the script(s) to the bone(s)".

To undo, simply right click the script-name in your inspector-panel and select "Click here to remove the script(s) to the bone(s)".

You can also click the three small dots to the right of the script-name to open the same context menu as right clicking.